+++ src/components/portal/__internal__/portal.context.ts
@@ -1,0 +1,7 @@
+import React from "react";
+
+export interface PortalContextProps {
+  renderInRoot?: boolean;
+}
+
+export default React.createContext<PortalContextProps>({});
